+++ packages/client/src/components/list/index.ts
@@ -1,2 +1,4 @@
 export * from './Card';
+export * from './CardFavorite';
+export * from './CardTags';
 export * from './List';